+++ NEWS.md
@@ -1,6 +1,13 @@
+# dhis2r 0.2.0
+
+## Improvements 
+
+* Add authentication via personal access token @Brar
+
+* Added   `get_any_analytic()` to cater for long DHIS2 favorites 
+
 # dhis2r 0.1.1
-Updated tests for analytics to be flexible to cover unavailable data on a DHIS2 instance
 
-# dhis2r 0.1.0
+* Fix R CMD check failures on CRAN
 
-* Added a `NEWS.md` file to track changes to the package.
+* Added a NEWS.md file to track changes to the package.